+++ src/main/java/com/atsuishio/superbwarfare/item/gun/data/GunData.java
@@ -274,7 +274,7 @@ public class GunData {
         return defaultGunData().burstAmount;
     }
 
-    public double shootDelay() {
+    public int shootDelay() {
         return defaultGunData().shootDelay;
     }
